Refactor: Make it work
Remove the unused `structs` and `trait` and implement a `get_roots`
function straight on the `DType` enum - because apparently that's a
thing that can happen(!)

The trickiest thing was to understand how to extract (or maybe unwrap?)
the <i32> type out of the different DType variations.

The `match` statement not only can allow for nested matches (e.g.
`DType::DPositive`), but also can bind the matched value to a variable
(e.g. `DType::DPositive(d)`) which allows us to use it easier in the
relevant `match` block.